Move FeedItemsMapper to its own file
